+++ com.unity.perception/Runtime/GroundTruth/SimulationState.cs
@@ -66,7 +66,6 @@ namespace UnityEngine.Perception.GroundTruth
         const float k_SimulationTimingAccuracy = 0.01f;
         const int k_MinPendingCapturesBeforeWrite = 150;
         const int k_MinPendingMetricsBeforeWrite = 150;
-        const float k_MaxDeltaTime = 100f;
 
         public SimulationState(string outputDirectory)
         {
@@ -473,7 +472,7 @@ namespace UnityEngine.Perception.GroundTruth
             }
 
             //find the deltatime required to land on the next active sensor that needs simulation
-            float nextFrameDt = k_MaxDeltaTime;
+            var nextFrameDt = float.PositiveInfinity;
             foreach (var activeSensor in m_ActiveSensors)
             {
                 float thisSensorNextFrameDt = -1;
@@ -491,10 +490,12 @@ namespace UnityEngine.Perception.GroundTruth
                 }
 
                 if (thisSensorNextFrameDt > 0f && thisSensorNextFrameDt < nextFrameDt)
+                {
                     nextFrameDt = thisSensorNextFrameDt;
+                }
             }
 
-            if (Math.Abs(nextFrameDt - k_MaxDeltaTime) < 0.0001)
+            if (float.IsPositiveInfinity(nextFrameDt))
             {
                 //means no sensor is controlling simulation timing, so we set Time.captureDeltaTime to 0 (default) which means the setting does not do anything
                 nextFrameDt = 0;