+++ com.unity.perception/Runtime/GroundTruth/SimulationState.cs
@@ -66,7 +66,6 @@ namespace UnityEngine.Perception.GroundTruth
         const float k_SimulationTimingAccuracy = 0.01f;
         const int k_MinPendingCapturesBeforeWrite = 150;
         const int k_MinPendingMetricsBeforeWrite = 150;
-        const float k_MaxDeltaTime = 100f;
 
         public SimulationState(string outputDirectory)
         {
@@ -447,6 +446,18 @@ namespace UnityEngine.Perception.GroundTruth
             {
                 var sensorData = m_Sensors[activeSensor];
 
+#if UNITY_EDITOR
+                if (UnityEditor.EditorApplication.isPaused)
+                {
+                    //When the user clicks the 'step' button in the editor, frames will always progress at .02 seconds per step.
+                    //In this case, just run all sensors each frame to allow for debugging
+                    Debug.Log($"Frame step forced all sensors to synchronize, changing frame timings.");
+
+                    sensorData.sequenceTimeOfNextRender = UnscaledSequenceTime;
+                    sensorData.sequenceTimeOfNextCapture = UnscaledSequenceTime;
+                }
+#endif
+
                 if (Mathf.Abs(sensorData.sequenceTimeOfNextRender - UnscaledSequenceTime) < k_SimulationTimingAccuracy)
                 {
                     //means this frame fulfills this sensor's simulation time requirements, we can move target to next frame.
@@ -460,6 +471,8 @@ namespace UnityEngine.Perception.GroundTruth
                         sensorData.sequenceTimeOfNextCapture += sensorData.renderingDeltaTime * (sensorData.framesBetweenCaptures + 1);
                         Debug.Assert(sensorData.sequenceTimeOfNextCapture > UnscaledSequenceTime,
                             $"Next scheduled capture should be after {UnscaledSequenceTime} but is {sensorData.sequenceTimeOfNextCapture}");
+                        while (sensorData.sequenceTimeOfNextCapture <= UnscaledSequenceTime)
+                            sensorData.sequenceTimeOfNextCapture += sensorData.renderingDeltaTime * (sensorData.framesBetweenCaptures + 1);
                     }
                     else if (sensorData.captureTriggerMode.Equals(CaptureTriggerMode.Manual))
                     {
@@ -473,7 +486,7 @@ namespace UnityEngine.Perception.GroundTruth
             }
 
             //find the deltatime required to land on the next active sensor that needs simulation
-            float nextFrameDt = k_MaxDeltaTime;
+            var nextFrameDt = float.PositiveInfinity;
             foreach (var activeSensor in m_ActiveSensors)
             {
                 float thisSensorNextFrameDt = -1;
@@ -491,10 +504,12 @@ namespace UnityEngine.Perception.GroundTruth
                 }
 
                 if (thisSensorNextFrameDt > 0f && thisSensorNextFrameDt < nextFrameDt)
+                {
                     nextFrameDt = thisSensorNextFrameDt;
+                }
             }
 
-            if (Math.Abs(nextFrameDt - k_MaxDeltaTime) < 0.0001)
+            if (float.IsPositiveInfinity(nextFrameDt))
             {
                 //means no sensor is controlling simulation timing, so we set Time.captureDeltaTime to 0 (default) which means the setting does not do anything
                 nextFrameDt = 0;